Fix remote viewer 'Service not found' by using uint16_t for port
Default port 49000 was passed as short (max 32767), truncating to -16536
and causing resolver to fail. Use uint16_t for port in NetworkClient and
tsdRemoteViewer; clamp UI port input to 1-65535. Update server to use
uint16_t for consistency.

+++ tsd/apps/interactive/demos/network/server/RenderServer.cpp
@@ -27,7 +27,7 @@ RenderServer::RenderServer(int argc, const char **argv)
 
 RenderServer::~RenderServer() = default;
 
-void RenderServer::run(short port)
+void RenderServer::run(uint32_t port)
 {
   m_port = port;
